separated mix tasks to their own module files, still need to DRY them up and fix documentation
